backend/pyproject.toml
[project]
name = "historai-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "groq",
    "python-dotenv",
    "supabase",
    "pydantic",
]